docs: Move IntentExample from IntentOpener folder to docs folder
